steamcompmgr: Move send_done to just before latest latch time
Makes more sense instead of sending done after we commit for page flip, otherwise the cadence can be slightly too ahead.

(Accounts for the bubble of time after latch -> commit being included in the time when we want to submit)

This also ensures that in the case where QueuePresent can stall on Wayland WSI (ew, gross!) that that stall will be finished before the next acquire.