prefer DRM formats over VkFormat
DRMFormatToVulkan is lossy; it doesn't distinguish between e.g. XR30 and
AR30. The primary plane on my laptop (8086:5917) supports XR30 but not
AR30, so when we get mixed up and try to use AR30 it fails and we crash.
Deal in DRM formats instead to avoid having to carry around an extra
hasAlpha flag.

Fixes #1215, but we still crash on my laptop for other reasons, so it
looks like Intel needs more fixes.